requirement.txt
Django==2.2
django-braces==1.15.0
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
